Put data def and parsing code to proper places. `try_from()` and associated error types for string->enum mappings should go under format/ node data should all be defined under node/
