+++ inox2d/src/formats.rs
@@ -1,3 +1,4 @@
+mod enums;
 pub mod inp;
 mod json;
 mod payload;
