+++ inox2d/src/node.rs
@@ -3,9 +3,7 @@ pub mod tree;
 
 use std::fmt::Debug;
 
-use crate::math::transform::TransformOffset;
-
-use data::InoxData;
+use data::{InoxData, TransformOffset};
 
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
 #[repr(transparent)]
